% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{performance_metrics}
\alias{performance_metrics}
\title{Performance Metrics of Candidate Models}
\format{
A data frame with 7 rows and 4 columns:
\describe{
\item{Indicator}{The name of the performance metric.}
\item{XGBoost}{Performance values for the XGBoost model.}
\item{RF (90\% Training)}{Performance values for the Random Forest model trained on 90\% of the data.}
\item{RF (80\% Training)}{Performance values for the Random Forest model trained on 80\% of the data.}
}
}
\usage{
performance_metrics
}
\description{
A dataset containing key performance indicators (e.g., AUROC, Sensitivity)
for the candidate models evaluated during package development.
}
\keyword{datasets}
