% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_rf_diagnostics}
\alias{plot_rf_diagnostics}
\title{Plot Diagnostic Charts for the Random Forest Model}
\usage{
plot_rf_diagnostics(newdata = NULL)
}
\arguments{
\item{newdata}{A data frame to evaluate for the confusion matrix. If NULL (the default),
the function uses the internal Out-of-Bag confusion matrix of the model.}
}
\value{
A combined ggplot object created by \code{grid.arrange}.
}
\description{
Displays the Out-of-Bag (OOB) error rate curve and a confusion matrix of the model.
}
\details{
The OOB error plot is static and reflects performance of the model during training.
The confusion matrix is dynamic: if newdata is provided, it shows performance
on that data; otherwise, it defaults to showing the performance on the OOB samples.
}
\examples{
\dontrun{
# This shows the OOB error and the OOB confusion matrix
plot_rf_diagnostics()

# This shows the OOB error and a new confusion matrix based on sample_variants
plot_rf_diagnostics(ConflictPredictR::sample_variants)
}
}
