% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predict_conflict_risk}
\alias{predict_conflict_risk}
\title{Predict the Risk of Conflicting Interpretations for Variants}
\usage{
predict_conflict_risk(newdata)
}
\arguments{
\item{newdata}{A data frame of new variants to predict.}
}
\value{
A tibble that includes the original data plus new columns for the
predicted conflict class (\code{.pred_conflict_class}) and the conflict risk
score (\code{.pred_conflict_risk}).
}
\description{
This function takes new variant data and returns a "Conflict Risk Score"
from the pre-trained Random Forest model.
}
\examples{
# Run predictions on the built-in sample data
risk_scores <- predict_conflict_risk(ConflictPredictR::sample_variants)
# View a subset of the results
print(head(risk_scores[, c("CADD_PHRED", ".pred_conflict_class", ".pred_conflict_risk")]))
}
