% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predict_pathogenicity}
\alias{predict_pathogenicity}
\title{Predict Variant Pathogenicity using Random Forest}
\usage{
predict_pathogenicity(newdata)
}
\arguments{
\item{newdata}{A data frame of new variants to predict.}
}
\value{
A tibble that includes the original data plus new columns for the predicted class and the probability of being pathogenic.
}
\description{
This function takes new variant data and returns predictions from the
pre-trained Random Forest models.
}
\examples{
# Run predictions on the built-in sample data
predictions <- predict_pathogenicity(PathoPrediction::sample_variants)
print(predictions)
}
